Address pydantic review comments
